Setup project, moved AccumuloIngestCommand into project.
