+++ .gitignore
@@ -1,0 +1,29 @@
+project/boot
+project/plugins/project
+project/plugins/target
+project/target
+target
+.ensime
+TAGS
+\#*#
+*~
+.#*
+.lib
+*.aux.xml
+*.jar
+*.crc
+_SUCCESS
+
+*.pyc
+.project
+.classpath
+.cache
+.settings
+.history
+.idea
+.DS_Store
+*.iml
+*.swp
+*.swo
+*.sublime-*
+.vagrant
