+++ server/src/main/resources/application.conf
@@ -1,0 +1,7 @@
+akka {
+  loglevel = INFO
+}
+
+spray.can.server {
+  request-timeout = 10s
+}
