+++ server/src/main/scala/geotrellis/admin/server/CorsSupport.scala
@@ -1,0 +1,18 @@
+package geotrellis.admin.server
+
+import spray.http.HttpHeaders._
+import spray.http.HttpMethods._
+import spray.http.{ AllOrigins, MediaTypes }
+import spray.http.{ HttpMethods, HttpMethod, HttpResponse, AllOrigins }
+import spray.routing._
+
+trait CorsSupport { self: HttpService =>
+  val corsHeaders = List(`Access-Control-Allow-Origin`(AllOrigins),
+    `Access-Control-Allow-Methods`(GET, POST, OPTIONS, DELETE),
+    `Access-Control-Allow-Headers`("Origin, X-Requested-With, Content-Type, Accept, Accept-Encoding, Accept-Language, Host, Referer, User-Agent, Access-Control-Request-Method, Access-Control-Request-Headers"))
+
+  def cors: Directive0 = {
+    val rh = implicitly[RejectionHandler]
+    respondWithHeaders(corsHeaders) & handleRejections(rh)
+  }
+}
