--- service/.gitignore
@@ -1,8 +1,0 @@
-/RUNNING_PID
-/logs/
-/project/*-shim.sbt
-/project/project/
-/project/target/
-/target/
-target
-.idea/
