--- service/README.md
@@ -1,1 +1,0 @@
-# Catalog and TMS Service
