--- service/src/main/resources/application.conf
@@ -1,9 +1,0 @@
-akka {
-  loglevel = INFO
-}
-
-spray.can.server {
-  request-timeout = 10s
-}
-
-geotrellis.accumulo.catalog = "metadata_fossdem"
