--- service/src/main/scala/climate/rest/RouteExceptionHandlers.scala
@@ -1,16 +1,0 @@
-package climate.rest
-
-
-import spray.http.StatusCodes
-import spray.httpx.SprayJsonSupport._
-import spray.routing.Directive._
-import spray.routing.{ExceptionHandler, HttpService}
-import spray.util.LoggingContext
-
-trait RouteExceptionHandlers extends HttpService {
-  
-  implicit def exceptionHandler(implicit log: LoggingContext) = ExceptionHandler {
-    case e: Exception => complete(StatusCodes.NotFound, e.getMessage())        
-  }    
-
-}
